dev-0524 Moved StartTestCaseInfo and EndTestCaseInfo to a separate type file with limited message imports
