+++ cucumber-tsflow-specs/node/package.json
@@ -2,7 +2,7 @@
 	"private": true,
 	"name": "cucumber-tsflow-node",
 	"description": "Specification for 'cucumber-tsflow'.",
-	"version": "6.5.4",
+	"version": "6.5.5",
 	"author": "Lonnie Wall <lynxdev@lynxwall.com>",
 	"license": "MIT",
 	"scripts": {
