+++ cucumber-tsflow/src/cucumber/managed-scenario-context.ts
@@ -2,7 +2,7 @@ import _ from 'underscore';
 
 import { ContextType } from '../types/types';
 import { ScenarioContext, ScenarioInfo } from '../types/scenario-context';
-import { EndTestCaseInfo, StartTestCaseInfo } from './message-collector';
+import { EndTestCaseInfo, StartTestCaseInfo } from '../types/test-case-info';
 import { World } from '@cucumber/cucumber';
 
 class ActiveInfo {
