+++ cucumber-tsflow/src/cucumber/test-case-runner.ts
@@ -20,7 +20,7 @@ import { BindingRegistry } from './binding-registry';
 import { StepBinding } from '../types/step-binding';
 import { ManagedScenarioContext } from './managed-scenario-context';
 import { error } from 'console';
-import { EndTestCaseInfo, StartTestCaseInfo } from './message-collector';
+import { EndTestCaseInfo, StartTestCaseInfo } from '../types/test-case-info';
 
 export default class TestCaseRunner {
 	private readonly attachmentManager: AttachmentManager;
