+++ cucumber-tsflow/src/index.ts
@@ -2,3 +2,4 @@ export * from './cucumber/binding-decorator';
 export * from './cucumber/hook-decorators';
 export * from './cucumber/step-definition-decorators';
 export { ScenarioContext, ScenarioInfo } from './types/scenario-context';
+export { StartTestCaseInfo, EndTestCaseInfo } from './types/test-case-info';
