+++ cucumber-tsflow/src/version.ts
@@ -1,2 +1,2 @@
 // Generated by genversion.
-export const version = '6.5.4'
+export const version = '6.5.5'
